This merge request merges the changes from the ~"21.3" branch between [release/21.3.13](https://gitlab.cern.ch/atlas/athena/tags/release/21.3.13)
and [release/21.3.14](https://gitlab.cern.ch/atlas/athena/tags/release/21.3.14) (including the changes from the ~"21.0"
branch between [release/21.0.95](https://gitlab.cern.ch/atlas/athena/tags/release/21.0.95)
and [release/21.0.97](https://gitlab.cern.ch/atlas/athena/tags/release/21.0.97)) into the ~"21.9" branch.
This is done via a merge of [release/21.3.14](https://gitlab.cern.ch/atlas/athena/tags/release/21.3.14).
(Last ~"21.3" release to be merged into ~"21.9" was
[release/21.3.13](https://gitlab.cern.ch/atlas/athena/tags/release/21.3.13).)

[release/21.0.96](https://gitlab.cern.ch/atlas/athena/tags/release/21.0.96) is a Tier0 release. It alters the output of heavy ion production jobs.
[release/21.0.97](https://gitlab.cern.ch/atlas/athena/tags/release/21.0.97) is a Tier0 release. It is FT0-preserving.
[release/21.3.14](https://gitlab.cern.ch/atlas/athena/tags/release/21.3.14) is otherwise FT0-preserving.

Here is the full diff between https://gitlab.cern.ch/atlas/athena/compare/release/21.3.13...release/21.3.14

NB The changes in !22913 (which were part of [release/21.0.96](https://gitlab.cern.ch/atlas/athena/tags/release/21.0.96)). Have been
left out of this merge due to conflicts with the existing code in ~"21.9".

+++ Trigger/TrigT1/TrigT1Interfaces/TrigT1Interfaces/RecJetRoI.h
@@ -109,7 +109,7 @@ namespace LVL1 {
       unsigned int windowSize( unsigned int thresh ) const;
       
       /** returns bitmask of passed thresholds */
-      unsigned int thresholdPattern() const;
+      unsigned long int thresholdPattern() const;
 
       /** returns true if thresh is a valid threshold number */
       bool isValidThreshold( unsigned int thresh ) const;
